+++ src/routes/api/deleteBox/+server.ts
@@ -1,4 +1,6 @@
+import { json } from '@sveltejs/kit';
 import { MongoClient, ServerApiVersion } from 'mongodb';
+import type { RequestHandler } from './$types';
 import dotenv from 'dotenv';
 dotenv.config();
 
@@ -17,32 +19,25 @@ const client = new MongoClient(uri, {
 	}
 });
 
-/** @type {import('./$types').PageLoad} */
-export async function load({ params }) {
-	const id = (params.slug);
-	let contents, images, boxExist;
+export const POST: RequestHandler = async ({ request }) => {
+	const { id } = await request.json();
+	let res;
+
 	try {
 		// Connect the client to the server	(optional starting in v4.7)
 		await client.connect();
 		const collection = client.db('test-box-db').collection('boxes');
 
-		let box = await collection.findOne(
-			{ _id: id },
-			{ sort: { lastModified: -1 }, projection: {} }
-		);
-
-		boxExist = box != null;
-		contents = boxExist ? box.contents : '';
-		images = boxExist ? box.images : [];
+		res = await collection.deleteOne({ _id: id });
 	} finally {
 		// Ensures that the client will close when you finish/error
-		if (client) await client.close();
+		await client.close();
 	}
-
-	return {
-		box: id,
-		contents,
-		images,
-		boxExist
-	};
-}
+	if (res.deletedCount === 0) {
+		return json({ error: 'No box with id found' }, { status: 404 });
+	} else if (res.deletedCount === 1) {
+		return json({ status: 'ok' });
+	} else {
+		return json({ error: 'Unexpected Server Error' }, { status: 500 });
+	}
+};
