+++ src/routes/api/saveImage/+server.ts
@@ -1,7 +1,8 @@
 import { json } from '@sveltejs/kit';
 import { MongoClient, ServerApiVersion } from 'mongodb';
+import type { RequestHandler } from './$types';
 import dotenv from 'dotenv';
-import { fuzzyFilter } from 'fuzzbunny';
+import sharp from 'sharp';
 dotenv.config();
 
 let uri;
@@ -19,25 +20,33 @@ const client = new MongoClient(uri, {
 	}
 });
 
-/** @type {import('./$types').RequestHandler} */
-export async function GET({ url }) {
-	const query = decodeURIComponent(url.searchParams.get('query'));
-	let contents;
+export const POST: RequestHandler = async ({ request }) => {
+	const { id, base64 } = await request.json();
+
+	const base64Data = base64.replace(/^data:image\/\w+;base64,/, '');
+	const imageBuffer = Buffer.from(base64Data, 'base64');
+
+	const compressedBuffer = await sharp(imageBuffer)
+		.withMetadata()
+		.resize(6500, 6500, { fit: 'inside' })
+		.jpeg({ quality: 75 })
+		.toBuffer();
+
+	const compressedBase64 = `data:image/jpeg;base64,${compressedBuffer.toString('base64')}`;
+
 	let res;
 	try {
 		// Connect the client to the server	(optional starting in v4.7)
 		await client.connect();
 		const collection = client.db('test-box-db').collection('boxes');
 
-		let contentsCursor = collection
-			.find({}, { projection: { contents: 1, images: 1 } })
-		contents = await contentsCursor.toArray();
+		res = await collection.updateOne(
+			{ _id: id },
+			{ $push: { images: compressedBase64 }, $set: { lastModified: Date.now() } }
+		);
 	} finally {
 		// Ensures that the client will close when you finish/error
 		await client.close();
 	}
-
-	res = fuzzyFilter(contents, query, { fields: ['_id', 'contents'] });
-
 	return json(res);
-}
+};
