+++ src/routes/api/search/1/+server.ts
@@ -1,5 +1,8 @@
+import { json } from '@sveltejs/kit';
 import { MongoClient, ServerApiVersion } from 'mongodb';
+import type { RequestHandler } from './$types';
 import dotenv from 'dotenv';
+import { fuzzyFilter } from 'fuzzbunny';
 dotenv.config();
 
 let uri;
@@ -17,34 +20,23 @@ const client = new MongoClient(uri, {
 	}
 });
 
-/** @type {import('./$types').PageLoad} */
-export async function load({ }) {
-	let lastPage;
+export const GET: RequestHandler = async ({ url }) => {
+	const query = decodeURIComponent(url.searchParams.get('query'));
 	let contents;
+	let res;
 	try {
 		// Connect the client to the server	(optional starting in v4.7)
 		await client.connect();
 		const collection = client.db('test-box-db').collection('boxes');
 
-		lastPage = Math.ceil((await collection.countDocuments()) / 10);
-
-		let contentsCursor = collection
-			.find({}, { sort: { lastModified: -1 }, projection: { images: 1, contents: 1 } })
-			.limit(10);
-
+		let contentsCursor = collection.find({}, { projection: { contents: 1, images: 1 } });
 		contents = await contentsCursor.toArray();
-		for (let i = 0; i < contents.length; i++) {
-			if (contents[i].images?.length > 5) {
-				contents[i].images = contents[i].images.slice(0, 5);
-			}
-		}
 	} finally {
 		// Ensures that the client will close when you finish/error
-		if (client) await client.close();
+		await client.close();
 	}
 
-	return {
-		contents: contents,
-		lastPage: lastPage
-	};
-}
+	res = fuzzyFilter(contents, query, { fields: ['_id', 'contents'] });
+
+	return json(res);
+};
