+++ src/routes/api/search/2/+server.ts
@@ -1,5 +1,6 @@
 import { json } from '@sveltejs/kit';
 import { MongoClient, ServerApiVersion } from 'mongodb';
+import type { RequestHandler } from './$types';
 import dotenv from 'dotenv';
 import { fuzzyFilter } from 'fuzzbunny';
 dotenv.config();
@@ -19,8 +20,7 @@ const client = new MongoClient(uri, {
 	}
 });
 
-/** @type {import('./$types').RequestHandler} */
-export async function GET({ url }) {
+export const GET: RequestHandler = ({ url }) => {
 	const query = decodeURIComponent(url.searchParams.get('query'));
 	let contents;
 	let res;
@@ -34,20 +34,19 @@ export async function GET({ url }) {
 		res = fuzzyFilter(contents, query, { fields: ['_id', 'contents'] });
 		//fetch imgs for each result _id as promise and append to json as they resolve
 		//await all promises
-		let dbRes = []
+		let dbRes = [];
 		for (let box of res) {
-			dbRes.push(collection.findOne({ _id: box.item._id }, { projection: { images: 1 } }))
+			dbRes.push(collection.findOne({ _id: box.item._id }, { projection: { images: 1 } }));
 		}
-		await Promise.all(dbRes)
-			.then((imgs) => {
-				for (let i = 0; i < dbRes.length; i++) {
-					if (imgs[i].images) res[i].item.images = imgs[i].images
-				}
-			})
+		await Promise.all(dbRes).then((imgs) => {
+			for (let i = 0; i < dbRes.length; i++) {
+				if (imgs[i].images) res[i].item.images = imgs[i].images;
+			}
+		});
 	} finally {
 		// Ensures that the client will close when you finish/error
 		await client.close();
 	}
 
 	return json(res);
-}
+};
