+++ src/routes/page/[slug]/+page.server.ts
@@ -1,4 +1,5 @@
 import { MongoClient, ServerApiVersion } from 'mongodb';
+import type { PageLoad } from './$types';
 import dotenv from 'dotenv';
 import { redirect } from '@sveltejs/kit';
 dotenv.config();
@@ -18,8 +19,7 @@ const client = new MongoClient(uri, {
 	}
 });
 
-/** @type {import('./$types').PageLoad} */
-export async function load({ params }) {
+export const load: PageLoad = async ({ params }) => {
 	//check if the page is a number, if not redirect to home
 	const regex = /^[0-9]+$/;
 	if (!regex.test(params.slug)) redirect(307, `/`);
@@ -61,4 +61,4 @@ export async function load({ params }) {
 		page: page,
 		lastPage: lastPage
 	};
-}
+};
